+++ frontend/app/src/components/accounts/manual-balances/ManualBalancesForm.vue
@@ -131,6 +131,7 @@ const save = async () => {
     const errorMessages = deserializeApiErrorMessage(status.message);
     if (errorMessages) {
       set(errors, (errorMessages?.balances[0] as any) ?? {});
+      await get(v$).$validate();
     } else {
       const obj = { message: status.message };
       setMessage({
@@ -165,6 +166,8 @@ watch(label, label => {
     const { label, ...data } = get(errors);
     set(errors, data);
   }
+
+  get(v$).label.$validate();
 });
 
 const customAssetFormRef: Ref<InstanceType<typeof CustomAssetForm> | null> =
@@ -250,7 +253,6 @@ onMounted(async () => {
     await form?.searchAssetPrice(editPayload.asset);
   }
 });
-
 defineExpose({
   save
 });