+++ frontend/app/src/components/history/LedgerActionForm.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import useVuelidate from '@vuelidate/core';
-import { helpers, required } from '@vuelidate/validators';
+import { helpers, required, requiredIf } from '@vuelidate/validators';
 import dayjs from 'dayjs';
 import { type PropType } from 'vue';
 import LocationSelector from '@/components/helper/LocationSelector.vue';
@@ -82,6 +82,18 @@ const rules = {
       t('ledger_action_form.location.validation.non_empty').toString(),
       required
     )
+  },
+  rate: {
+    required: helpers.withMessage(
+      t('ledger_action_form.rate.validation.non_empty').toString(),
+      requiredIf(refIsTruthy(rateAsset))
+    )
+  },
+  rateAsset: {
+    required: helpers.withMessage(
+      t('ledger_action_form.rate_asset.validation.non_empty').toString(),
+      requiredIf(refIsTruthy(rate))
+    )
   }
 };
 
@@ -90,7 +102,9 @@ const v$ = useVuelidate(
   {
     amount,
     asset,
-    location
+    location,
+    rate,
+    rateAsset
   },
   { $autoDirty: true, $externalResults: errorMessages }
 );
@@ -163,6 +177,8 @@ const save = async (): Promise<boolean> => {
       errorMessages,
       convertKeys(deserializeApiErrorMessage(result.message) ?? {}, true, false)
     );
+
+    await get(v$).$validate();
   }
 
   return false;
@@ -202,7 +218,6 @@ defineExpose({
       data-cy="location"
       :error-messages="v$.location.$errors.map(e => e.$message)"
       :label="tc('common.location')"
-      @focus="delete errorMessages['location']"
     />
 
     <date-time-picker
@@ -216,7 +231,6 @@ defineExpose({
       data-cy="datetime"
       :hint="tc('ledger_action_form.date.hint')"
       :error-messages="errorMessages['timestamp']"
-      @focus="delete errorMessages['timestamp']"
     />
 
     <v-row
@@ -234,7 +248,6 @@ defineExpose({
           required
           data-cy="asset"
           :error-messages="v$.asset.$errors.map(e => e.$message)"
-          @focus="delete errorMessages['asset']"
         />
       </v-col>
 
@@ -246,7 +259,6 @@ defineExpose({
           required
           data-cy="amount"
           :label="tc('common.amount')"
-          @focus="delete errorMessages['amount']"
         />
       </v-col>
 
@@ -261,7 +273,6 @@ defineExpose({
           required
           data-cy="action-type"
           :error-messages="errorMessages['actionType']"
-          @focus="delete errorMessages['actionType']"
         />
       </v-col>
     </v-row>
@@ -281,8 +292,7 @@ defineExpose({
           data-cy="rate"
           :hint="tc('ledger_action_form.rate.hint')"
           :label="tc('ledger_action_form.rate.label')"
-          :error-messages="errorMessages['rate']"
-          @focus="delete errorMessages['rate']"
+          :error-messages="v$.rate.$errors.map(e => e.$message)"
         />
       </v-col>
       <v-col cols="12" md="4">
@@ -293,8 +303,7 @@ defineExpose({
           :hint="tc('ledger_action_form.rate_asset.hint')"
           persistent-hint
           data-cy="rate-asset"
-          :error-messages="errorMessages['rateAsset']"
-          @focus="delete errorMessages['rateAsset']"
+          :error-messages="v$.rateAsset.$errors.map(e => e.$message)"
         />
       </v-col>
     </v-row>
@@ -308,7 +317,6 @@ defineExpose({
       :label="tc('ledger_action_form.link.label')"
       :hint="tc('ledger_action_form.link.hint')"
       :error-messages="errorMessages['link']"
-      @focus="delete errorMessages['link']"
     />
 
     <v-textarea
@@ -320,7 +328,6 @@ defineExpose({
       :label="tc('ledger_action_form.notes.label')"
       :hint="tc('ledger_action_form.notes.hint')"
       :error-messages="errorMessages['notes']"
-      @focus="delete errorMessages['notes']"
     />
   </v-form>
 </template>