+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :bug:`-` Fixes bug where error message is not showed when some input filled with incorrect value.
 * :bug:`-` Fixes bug where Maker Vault Collateralization Ratio can't be edited in watcher form.
 * :bug:`-` Fixes bug where wrong account label was used in asset location breakdown.
 * :bug:`5560` EVM transaction events should now be properly ignored in accounting.